+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   * Learn from a Companion
   * Random Encounter
   * Stopwatch
+  * Companion Effects to Chat
 
 ## [0.1.0] - 2023-06-11